+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - `as.latex()` gains a `cite` argument to include citation in the results
 - Collected quotes vignette added for complete overview of quotes collection
+- `cite=TRUE` is now the default in `as.markdown()`, `as.latex()`, and  `search_quotes()`
 
 # Version 0.3.1 (2022-10-10)